vendor/xilinx: meta-xilinx-standalone-experimental: open-amp-xlnx: Add support for FreeRTOS
Enable builds with FreeRTOS OS in addition to Standalone for supported demos.

Since this update is pulling in from latest sources, clean up and update
OpenAMP YAML's to align with latest oo.

As part of this patch also clean up yocto task orderings to all be in section
at bottom of file for readability.

Signed-off-by: Ben Levinsky <ben.levinsky@amd.com>
Signed-off-by: Mark Hatle <mark.hatle@amd.com>